requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.6.3
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.0
cfscrape==2.1.1
charset-normalizer==2.0.12
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
cryptography==37.0.2
Django==4.2.6
exceptiongroup==1.1.3
fake-useragent==0.1.11
Flask==3.0.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.9
lxml==4.9.3
MarkupSafe==2.1.3
outcome==1.2.0
pip==23.3
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-decouple==3.8
requests==2.28.0
requests-toolbelt==1.0.0
selenium==4.2.0
selenium-wire==4.6.4
setuptools==58.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.9
Werkzeug==3.0.0
wsproto==1.1.0
zipp==3.17.0
zstandard==0.17.0